test: add stories for edges
